refactor: Moved suface_ops.rs out of surface/ folder.
